bonus card

+++ src/main/java/ru/project/mobile/dto/OrderDto.java
@@ -8,6 +8,6 @@ import java.util.Map;
 
 @Data
 public class OrderDto {
-    private final Map<String, Double> bookList;
+    private final Map<String, Integer> bookList;
     private final double fullPrice;
 }
